% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nest.R
\name{make_nest}
\alias{make_nest}
\title{Nest flat data given a template.}
\usage{
make_nest(x, template, prefix = "")
}
\arguments{
\item{x}{a character vector of data that matches keys in a template}

\item{template}{a nested list with keys corresponding to matching data}

\item{prefix}{a prefix to add to the template keys to match the data in x}
}
\value{
the template filled with the matching keys in x
}
\description{
Nest flat data given a template.
}
\details{
This is a toy implementation of Nested Template Data:
\url{https://rosettacode.org/wiki/Nested_templated_data#R}
}
\examples{
if (requireNamespace("jsonlite")) {
  # You can use JSON data
  template <- jsonlite::read_json(system.file("extdata", "template1.json", package = "grosbeak"))
} else {
  template <- list(list(c(1, 2), c(3, 4), 5))
}
payload  <- paste0("Payload#", 0:6)
make_nest(payload, template, prefix = "Payload#") 
}
