% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renest.R
\name{renest}
\alias{renest}
\title{Nest flat data given a template.}
\usage{
renest(x, template, prefix = "")
}
\arguments{
\item{x}{a character vector of data that matches keys in a template}

\item{template}{a nested list with keys corresponding to matching data}

\item{prefix}{a prefix to add to the template keys to match the data in x}
}
\value{
the template filled with the matching keys in x
}
\description{
Nest flat data given a template.
}
\references{
\url{https://rosettacode.org/wiki/Nested_templated_data#R}

if (requireNamespace("jsonlite")) {You can use JSON data

template <- read_json(system.file("extdata", "example1.json"))
} else {
template <- list(list(c(1, 2), c(3, 4), 5))
}
payload  <- paste0("Payload#", 0:6)
renest(payload, template, prefix = "Payload#")
}
